+++ packages/moonstone/Panels/ApplicationCloseButton.js
@@ -23,6 +23,15 @@ const ApplicationCloseButton = kind({
 
 	propTypes: /** @lends moonstone/Panels.ApplicationCloseButton.prototype */ {
 		/**
+		* Sets the hint string read when focusing the application close button.
+		*
+		* @type {String}
+		* @default 'Exit app'
+		* @public
+		*/
+		'aria-label': PropTypes.string,
+
+		/**
 		 * The background-color opacity of this button; valid values are `'opaque'`, `'translucent'`,
 		 * `'lightTranslucent'` and `'transparent'`.
 		 *
@@ -49,11 +58,15 @@ const ApplicationCloseButton = kind({
 		className: 'applicationCloseButton'
 	},
 
-	render: ({backgroundOpacity, onApplicationClose, ...rest}) => {
+	computed: {
+		exitAppAriaLabel: ({'aria-label': ariaLabel}) => (ariaLabel == null) ? $L('Exit app') : ariaLabel
+	},
+
+	render: ({backgroundOpacity, exitAppAriaLabel, onApplicationClose, ...rest}) => {
 		return (
 			<IconButton
 				{...rest}
-				aria-label={$L('Exit app')}
+				aria-label={exitAppAriaLabel}
 				backgroundOpacity={backgroundOpacity}
 				onTap={onApplicationClose}
 				small