+++ packages/moonstone/Panels/Panels.js
@@ -3,6 +3,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {shape} from '@enact/ui/ViewManager';
 
+import $L from '../internal/$L';
 import IdProvider from '../internal/IdProvider';
 import Skinnable from '../Skinnable';
 
@@ -68,6 +69,15 @@ const PanelsBase = kind({
 		 * @public
 		 */
 		closeButtonBackgroundOpacity: PropTypes.oneOf(['opaque', 'translucent', 'lightTranslucent', 'transparent']),
+
+		/**
+		* Sets the hint string read when focusing the application close button.
+		*
+		* @type {String}
+		* @default 'Exit app'
+		* @public
+		*/
+		exitAppAriaLabel: PropTypes.string,
 
 		/**
 		 * Unique identifier for the Panels instance
@@ -137,12 +147,14 @@ const PanelsBase = kind({
 		className: ({noCloseButton, styler}) => styler.append({
 			hasCloseButton: !noCloseButton
 		}),
-		applicationCloseButton: ({closeButtonBackgroundOpacity, id, noCloseButton, onApplicationClose}) => {
+		exitAppAriaLabel: ({exitAppAriaLabel}) => (exitAppAriaLabel == null) ? $L('Exit app') : exitAppAriaLabel,
+		applicationCloseButton: ({closeButtonBackgroundOpacity, exitAppAriaLabel, id, noCloseButton, onApplicationClose}) => {
 			if (!noCloseButton) {
 				const closeId = id ? `${id}_close` : null;
 
 				return (
 					<ApplicationCloseButton
+						aria-label={exitAppAriaLabel}
 						backgroundOpacity={closeButtonBackgroundOpacity}
 						className={css.close}
 						id={closeId}
@@ -172,6 +184,7 @@ const PanelsBase = kind({
 
 	render: ({noAnimation, arranger, childProps, children, generateId, index, applicationCloseButton, ...rest}) => {
 		delete rest.closeButtonBackgroundOpacity;
+		delete rest.exitAppAriaLabel;
 		delete rest.noCloseButton;
 		delete rest.onApplicationClose;
 		delete rest.onBack;